+++ app/Http/Controllers/Admin/UserController.php
@@ -11,6 +11,7 @@ use App\Models\UserInfo;
 use Exeption;
 use App\Mail\SendMailUser;
 use Mail;
+use DB;
 
 class UserController extends Controller
 {
@@ -141,9 +142,15 @@ class UserController extends Controller
     public function destroy(User $user)
     {
         if (!$user->role == User::ADMIN_ROLE) {
+            DB::beginTransaction();
             try {
+                $user->comments()->delete();
+                $user->posts()->delete();
+                $user->orders()->delete();
                 $user->delete();
+                DB::commit();
             } catch (Exception $e) {
+                DB::rollback();
                 session()->flash('message', trans('messages.delete_user_fail'));
             }
             return redirect()->route('admin.users.index')->with('message', trans('messages.delete_user_success'));
